requirements.txt
# Required Python packages

numpy
PyQt5
pytest
pytest-qt
pyzmq
psycopg2
simple_pid
tinkerforge
